Refactor design and implementation smells
