+++ glitch/analysis/design/__init__.py
@@ -1,0 +1,7 @@
+import os
+from typing import List
+
+__all__: List[str] = []
+for file in os.listdir(os.path.dirname(__file__)):
+    if file.endswith(".py") and file != "__init__.py" and file != "visitor.py":
+        __all__.append(file[:-3])  # type: ignore
